Simply CMS page to just a rich text field.
Also move "basic" html/css for page content block into base template
so that there is less need to duplicate this basic page setup html
in all rendered page templates. If the basic setup for a page is not
appropriate for a page then that page can override full-content block
rather than content.