Remove non-functional uservoice.js
I'm not sure why we have this since so far as I know the link
to uservoice on this site will simply be links, nothing programmatic.
As-is it is causing a javascript error as it's assuming something has
provided a conditionalLoad function and we don't have whatever that is
in place. Until we hear some javascript for uservoice is needed it's
best to simply remove it.